requirements.txt
# requirements.txt

numpy
pandas
scikit-learn
xgboost
lightgbm
catboost
optuna
hyperopt
PyYAML
prettytable
seaborn
matplotlib
pytest
mlflow
kaggle
requests